GEODE-8705 - Create modular JarDeploymentService implementation. 	   - Add ModularClasspathService 	   - Add testcontainers for testing in acceptance test            - Add verification when deploying jars with the "geode-" in the path.
GEODE-8705 - Adding ability to run DUnits with modularity enabled

+++ geode-cq/src/distributedTest/java/org/apache/geode/cache/query/cq/dunit/CqQueryUsingPoolDUnitTest.java
@@ -121,7 +121,7 @@ public class CqQueryUsingPoolDUnitTest extends JUnit4CacheTestCase {
 
   private static final int WAIT_DEFAULT = (20 * 1000);
 
-  public static final long MAX_TIME = Integer.getInteger(WAIT_PROPERTY, WAIT_DEFAULT).intValue();
+  public static final long MAX_TIME = Integer.getInteger(WAIT_PROPERTY, WAIT_DEFAULT);
 
   public final String[] cqs = new String[] {
       // 0 - Test for ">"
